Create Generator application
This application is generating "fake" data for the sake of this
talk. It is implementing all remote services that our dasboard
application is calling.

For the rest of this talk, we'll just consider that service as a
black box, something out of our control, and focus instead on the
dashboard application itself.