Dashboard application with Spring Boot 1.5.x
This Spring Boot 1.5.x application is the existing application
we'd like to upgrade to Spring Boot 2.0.x (and then, 2.1.x).

It is displaying charts showing the number of projects created by
start.spring.io over a period, and the events that occured during
that timeline. It is also fetching the top client IPs to detect
spam and abuse.

The application is currently using Spring MVC and Spring Security,
with Spring Boot Actuator as well.