Upgrade to Spring Boot 2.0.x
This commit upgrades the application to Spring Boot 2.0.x by changing
the Spring Boot parent version in the POM.

We can also update the POM to:

* directly depend on starter-web, since starter-thymeleaf does not
depend on it anymore (it works with both Spring MVC and WebFlux)
* `webjars-locator-core` is now the preferred dependency for WebJars
support
* the new json-starter already brings common Jackson modules, including
the JSR310 support module

Spring Framework deprecated `WebMvcConfigurerAdapter` in favor of
default implementations in the `WebMvcConfigurer` interface.

The application now compiles and runs without warnings, but it still
doesn't work as it used to.