Add timeout support on StatsClient
Reactor operators can help you implement advanced strategies on
asynchronous operations. Here, we'll wait until the reverse IP
lookup call completes or hits a timeout that we've configured.
When hit, we can use a fallback `Publisher` to do the work.

Such an operation would be harder to implement with imperative
APIs.