Add Tanzu Observability by Wavefront starter
This commit adds the Wavefront Starter to emit metrics to a Tanzu
Observaiblity backend. By default, the starter creates a freemium
account automatically.

To help identifying our app we add the name of the application and the
name of the service. Because this is a demo, we also increase the
frequencey at which Micrometer ships metrics to Wavefront.